Split out Clarity character stat details
