+++ src/app/store-stats/ClarityCharacterStat.m.scss.d.ts
@@ -1,13 +1,10 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'boostedValue': string;
-  'breakdown': string;
-  'breakdownValue': string;
-  'icon': string;
-  'label': string;
-  'row': string;
-  'values': string;
+  'bar': string;
+  'barCurrent': string;
+  'communityInsightSection': string;
+  'graph': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
